requirements/base.txt
Django==3.2.12
psycopg2==2.8.6
django-redis==5.0.0
celery==5.2.2
PyJWT==2.3.0
django-cors-headers==3.7.0
django-extensions==3.1.3
djangorestframework-simplejwt==4.8.0
django-softdelete==0.9.2
django-storages==1.11.1
boto3==1.17.93
requests==2.25.1
Pillow==9.0.1
djangorestframework==3.12.4
django-celery-beat==2.2.1
django-sendgrid-v5
click==8.0.4
